+++ hrk_nasa_apis_test/test/services/ssd_cneos/sbdb_cad_api/data/sbdb_cad_body/sbdb_cad_body_test.dart
@@ -28,6 +28,13 @@ void main() {
         expect(sbdbCadBody.data, null);
       });
 
+      test('200/total-only-zero', () {
+        final sbdbCadBody = SbdbCadBodyExt.getSample('200/total-only-zero');
+        expect(sbdbCadBody.count, 0);
+        expect(sbdbCadBody.total, 0);
+        expect(sbdbCadBody.data, null);
+      });
+
       test('200/1', () {
         final sbdbCadBody = SbdbCadBodyExt.getSample('200/1');
         expect(sbdbCadBody.count, greaterThan(0));